Improve data-structure/React arch flow
This is kind of a big one. The idea here is that prior to this I was
doing everything as one huge JSON because I thought I could get away
with it. It turns out I can only to a certain extent. React starts
re-rendering every component, even if it had nothing to do with the
update, because I’m piping the entire data structure and neither
React/Firebase do data-level diffs. I wanted to keep my old code,
though, so I worked around this with a clever solution: deconstruct and
reconstruct the same old huge JSON data structure at runtime to
preserve the convenience of the old data structure but obtaining the
benefits of a sharded one.