Fix ridiculous firebase array handling
An empty array is null in firebase, which totally fucks up the way I
expected to handle the data. Because of this, the code is probably
going to be littered with (this || some default data) unless I do
some post-firebase filtering while adds unnecessary complexity, ruining
most libraries like ReactFire.