Fix background tick processing
So basically _.extend is crazy and mutates the original object
as opposed to returns a copy of the element? I wasn't even sure
that was legal in JS. Also needed to pass the player key into
the object merging cause I wasn't in the mood to mess around
with closures/scoping problems.